pyproject.toml
[tool.poetry]
name = "Back-end"
version = "0.1.0"
description = ""
authors = ["Barham Farraj <barhamfarraj@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
requests = "^2.28.0"
django-environ = "^0.9.0"
djangorestframework-simplejwt = "^5.2.0"
psycopg2-binary = "^2.9.3"
whitenoise = "^6.2.0"
django-phonenumber-field = "^6.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
